debug

+++ System/CPU/Register.cpp
@@ -24,20 +24,20 @@ Register::Register() {
 std::string Register::toString() const {
     std::ostringstream oss;
     oss <<
-        "PC: 0x" << std::hex << static_cast<uint16_t>(PC) << " " <<
-        "SP: 0x" << std::hex << static_cast<uint8_t>(SP)  << " " <<
-        "A: 0x" << std::hex << static_cast<uint8_t>(A) << " " <<
-        "X: 0x" << std::hex << static_cast<uint8_t>(X) << " " <<
-        "Y: 0x" << std::hex << static_cast<uint8_t>(Y) << std::endl <<
+        "PC: 0x" << std::hex << static_cast<unsigned int>(PC) << " " <<
+        "SP: 0x" << std::hex << static_cast<unsigned int>(SP)  << " " <<
+        "A: 0x" << std::hex << static_cast<unsigned int>(A) << " " <<
+        "X: 0x" << std::hex << static_cast<unsigned int>(X) << " " <<
+        "Y: 0x" << std::hex << static_cast<unsigned int>(Y) << std::endl <<
         "Flags: " <<
-        "C: " << std::hex << static_cast<uint8_t>(Flags.C) << " " <<
-        "Z: " << std::hex << static_cast<uint8_t>(Flags.Z) << " " <<
-        "I: " << std::hex << static_cast<uint8_t>(Flags.I) << " " <<
-        "D: " << std::hex << static_cast<uint8_t>(Flags.D) << " " <<
-        "B: " << std::hex << static_cast<uint8_t>(Flags.B) << " " <<
-        "_: " << std::hex << static_cast<uint8_t>(Flags._) << " " <<
-        "V: " << std::hex << static_cast<uint8_t>(Flags.V) << " " <<
-        "N: " << std::hex << static_cast<uint8_t>(Flags.N) << " " <<
+        "C: " << std::hex << static_cast<unsigned int>(Flags.C) << " " <<
+        "Z: " << std::hex << static_cast<unsigned int>(Flags.Z) << " " <<
+        "I: " << std::hex << static_cast<unsigned int>(Flags.I) << " " <<
+        "D: " << std::hex << static_cast<unsigned int>(Flags.D) << " " <<
+        "B: " << std::hex << static_cast<unsigned int>(Flags.B) << " " <<
+        "_: " << std::hex << static_cast<unsigned int>(Flags._) << " " <<
+        "V: " << std::hex << static_cast<unsigned int>(Flags.V) << " " <<
+        "N: " << std::hex << static_cast<unsigned int>(Flags.N) << " " <<
         std::endl;
     return oss.str();
 }
